requirements.txt
# Requirements pour le projet Hôpitaux Marocains
# Installation: pip install -r requirements.txt

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Database
pymysql>=1.1.0
sqlalchemy>=2.0.0

# PDF Processing (optionnel)
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Configuration
pyyaml>=6.0.0

# Logging et Utilitaires
python-dotenv>=1.0.0
tqdm>=4.65.0